Revert "refactor: make private rpc package to public and change timestamp format"
