+++ internal/rpc/coldport.go
@@ -19,7 +19,7 @@ type ColdPort int
 func (t *ColdPort) Push(ctx context.Context, args *LogMessage, reply *Reply) error {
 	if global.ColdRing.Size() >= global.DefaultRingSize {
 		reply.Rate = util.RangeMapping(global.ColdRing.Size())
-		return errors.New("cold port is full")
+		return errors.New("ColdPort is full!")
 	}
 	log := CopyLogMessage(args)
 	global.ColdRing.Push(log)
